made some int changes

+++ src/cs5004/animator/model/animation/Animator.java
@@ -28,17 +28,17 @@ public interface Animator extends ReadonlyAnimator {
   /**
    * Move the given shape to the new position.
    * @param name the name of the {@link Shape} to be moved, a String
-   * @param x new x coordinate, a double
-   * @param y new y coordinate, a double
-   * @param originalX old x coordinate, a double
-   * @param originalY old y coordinate, a double
+   * @param x new x coordinate, int
+   * @param y new y coordinate, int
+   * @param originalX old x coordinate, int
+   * @param originalY old y coordinate, int
    * @param start move start time, an int
    * @param stop move stop time, an int
    * @throws IllegalArgumentException if the start or stop times are out of bounds of the shape's
    *      appear/disappear window, if the stop time is less than or equal to the start time, or if
    *      the shape is already moving in this window, or if no shape if the list has the given name
    */
-  void move(String name, double x, double y, double originalX, double originalY, int start, int
+  void move(String name, int x, int y, int originalX, int originalY, int start, int
           stop) throws IllegalArgumentException;
 
   /**
@@ -64,16 +64,16 @@ public interface Animator extends ReadonlyAnimator {
    * Change the shape's scale.
    *
    * @param name name of the {@link Shape} to be scaled, a String
-   * @param width new width, a double
-   * @param height new height, a double
-   * @param originalHeight original height, a double
-   * @param originalWidth original width, a double
+   * @param width new width, int
+   * @param height new height, int
+   * @param originalHeight original height, int
+   * @param originalWidth original width, int
    * @param start scaling start time, an int
    * @param stop scaling stop time, an int
    * @throws IllegalArgumentException if the start or stop times are out of bounds of the shape's
    *      appear/disappear window, or if the shape's width and/or height is already scaling in this
    *      window, or if width and/or height <= 0, or if no shape in the list has the given name
    */
-  void scaleShape(String name, double width, double height, double originalHeight, double
+  void scaleShape(String name, int width, int height, int originalHeight, int
           originalWidth, int start, int stop) throws IllegalArgumentException;
 }
